Created base class for VR interface apps
